Harden native brand detection during plain cloning
Avoid consulting user-controlled toStringTag getters while retaining cross-realm recognition through intrinsic native accessors. Split the mixed regression into assertion-specific files so each new test has one named responsibility.

Constraint: Self-Review findings must be published before remediation

Rejected: Trust Object.prototype.toString for all routing | custom tags can demote valid natives and execute user getters

Confidence: high

Scope-risk: moderate

Directive: Keep intrinsic brand probes paired with real-native custom-tag and prototype-impostor regressions

Tested: focused plain_native_clone schema tests; scoped Prettier; git diff --check

Not-tested: repository-wide build and test remain pending after rebase
